updated the dashboard and server response for user. still need to fix jwt!
