+++ client/src/app/components/home/home-right/home-right.component.html
@@ -1,2 +1,1 @@
-<button md-button (click)="openDialog()">Launch dialog</button>
 <app-list></app-list>
