+++ client/src/app/components/search/filter-list/filter-list.component.html
@@ -51,7 +51,7 @@
     </ng-template>
 
     <button md-raised-button placement="bottom" #p1="ngbPopover" (document:click)="p1.close()" (click)="$event.stopPropagation();p2.close();p3.close();p4.close()"
-      [ngbPopover]="priceRange">Any Price <i class="material-icons">expand_more</i>
+      [ngbPopover]="priceRange">Price <i class="material-icons">expand_more</i>
 
     </button>
 
@@ -76,7 +76,7 @@
       </md-radio-group>
     </ng-template>
     <button md-raised-button placement="bottom" #p2="ngbPopover" (document:click)="p2.close()" (click)="$event.stopPropagation();p1.close();p3.close();p4.close()"
-      [ngbPopover]="beds">All Beds <i class="material-icons">expand_more</i>
+      [ngbPopover]="beds">Beds <i class="material-icons">expand_more</i>
     </button>
     <ng-template #propertyType>
       <section class="example-section" (click)="$event.stopPropagation()">
@@ -87,7 +87,7 @@
 
     </ng-template>
     <button md-raised-button placement="bottom" #p3="ngbPopover" (document:click)="p3.close()" (click)="$event.stopPropagation();p1.close();p2.close();p4.close()"
-      [ngbPopover]="propertyType">All Home Types <i class="material-icons">expand_more</i>
+      [ngbPopover]="propertyType">Home Types <i class="material-icons">expand_more</i>
     </button>
 
     Found: {{listingService.listingCount}} matches
